+++ sentry-kotlin-multiplatform/src/commonAppleMain/kotlin/io/sentry/kotlin/multiplatform/SentryBridge.apple.kt
@@ -44,7 +44,10 @@ internal actual object SentryBridge {
     }
 
     actual fun captureException(throwable: Throwable, scopeCallback: ScopeCallback): SentryId {
-        val cocoaSentryId = SentrySDK.captureException(throwable.asNSException(true), configureScopeCallback(scopeCallback))
+        val cocoaSentryId = SentrySDK.captureException(
+            throwable.asNSException(true),
+            configureScopeCallback(scopeCallback)
+        )
         return SentryId(cocoaSentryId.toString())
     }
 
@@ -89,8 +92,7 @@ internal actual object SentryBridge {
                 CocoaScopeProvider(it)
             }
             cocoaScopeProvider?.let {
-                val scope = Scope(it)
-                scopeCallback.invoke(scope)
+                scopeCallback.invoke(it)
             }
         }
     }