+++ sentry-kotlin-multiplatform/src/commonMain/kotlin/io/sentry/kotlin/multiplatform/SentryOptions.kt
@@ -3,8 +3,12 @@ package io.sentry.kotlin.multiplatform
 import io.sentry.kotlin.multiplatform.protocol.Breadcrumb
 import io.sentry.kotlin.multiplatform.protocol.SdkVersion
 
-public open class SentryOptions {
+private const val DEFAULT_MAX_BREADCRUMBS = 100
+private const val DEFAULT_MAX_ATTACHMENT_SIZE = 20 * 1024 * 1024L
+private const val DEFAULT_SESSION_INTERVAL_MILLIS = 30000L
 
+/** Sentry options that can be used to configure the SDK. */
+public open class SentryOptions {
     /**
      * The DSN tells the SDK where to send the events to. If this value is not provided, the SDK will
      * just not send any events.
@@ -40,14 +44,6 @@ public open class SentryOptions {
     /** Sets the distribution. Think about it together with release and environment */
     public var dist: String? = null
 
-    /**
-     * Configures the sample rate as a percentage of transactions to be sent in the range of 0.0 to 1.0.
-     * If set to 1.0 then 100% of transactions are sent.
-     * If set to 0.1 only 10% of transactions are sent.
-     * Transactions are picked randomly. Default is null (disabled)
-     */
-    public var tracesSampleRate: Double? = null
-
     /** This function is called by TracesSampler to determine if transaction is sampled - meant to be sent to Sentry. */
     public var tracesSampler: ((SamplingContext) -> Double?)? = null
 
@@ -58,7 +54,7 @@ public open class SentryOptions {
      * The session tracking interval in millis. This is the interval to end a session if the App goes
      * to the background.
      */
-    public var sessionTrackingIntervalMillis: Long = 30000
+    public var sessionTrackingIntervalMillis: Long = DEFAULT_SESSION_INTERVAL_MILLIS
 
     /**
      * Enables/Disables capturing screenshots before an error.
@@ -76,10 +72,10 @@ public open class SentryOptions {
     public var sdk: SdkVersion? = null
 
     /** This variable controls the total amount of breadcrumbs that should be captured. Default is 100. */
-    public var maxBreadcrumbs: Int = 100
+    public var maxBreadcrumbs: Int = DEFAULT_MAX_BREADCRUMBS
 
     /** This variable controls the max attachment size in bytes */
-    public var maxAttachmentSize: Long = 20 * 1024 * 1024
+    public var maxAttachmentSize: Long = DEFAULT_MAX_ATTACHMENT_SIZE
 
     /**
      * Enables or disables the attach view hierarchy feature when an error happened.
@@ -97,6 +93,7 @@ public open class SentryOptions {
 
     /**
      * A list of HTTP status code ranges indicating which client errors should be captured as errors.
+     *
      * By default, only HTTP client errors with a response code between 500 and 599 are captured as errors.
      *
      * Available on Apple.
@@ -104,10 +101,26 @@ public open class SentryOptions {
     public var failedRequestStatusCodes: List<HttpStatusCodeRange> = listOf(HttpStatusCodeRange())
 
     /**
-     * A list of HTTP request targets indicating which client errors should be captured as errors with either regex or a plain string.
+     * A list of HTTP request targets indicating which client errors should be captured as errors with
+     * either regex or a plain string.
+     *
      * By default, HTTP client errors from every target (.* regular expression) are automatically captured.
      *
      * Available on Apple.
      */
     public var failedRequestTargets: List<String> = listOf(".*")
+
+    /**
+     * Configures the sample rate as a percentage of events to be sent in the range of 0.0 to 1.0. if
+     * 1.0 is set it means that 100% of events are sent. If set to 0.1 only 10% of events will be
+     * sent. Events are picked randomly. Default is null (disabled)
+     */
+    public var sampleRate: Double? = null
+
+    /**
+     * Configures the sample rate as a percentage of transactions to be sent in the range of 0.0 to
+     * 1.0. if 1.0 is set it means that 100% of transactions are sent. If set to 0.1 only 10% of
+     * transactions will be sent. Transactions are picked randomly. Default is null (disabled)
+     */
+    public var tracesSampleRate: Double? = null
 }